+++ model/model.go
@@ -6,7 +6,7 @@ type (
 	//Log defines a log
 	Log struct {
 		ID        uint       `json:"id"`
-		Speed     int        `json:"speed"`
+		Speed     float32    `json:"speed"`
 		Dir       string     `json:"dir"` //direction
 		CreatedAt time.Time  `json:"created_at"`
 		UpdatedAt *time.Time `json:"updated_at"`
